requirements.txt
python-dotenv
pytest
pytest-playwright
